+++ front/src/components/CardsCatalog/CardsCatalog.css
