+++ front/src/components/ListEntry/ListEntry.css
@@ -1,4 +1,4 @@
-@import "../../../../assets/styles/elevation.css";
+@import "../../assets/styles/elevation.css";
 
 .list-entry {
   @mixin dp01;
