+++ front/src/components/SimpleCard/SimpleCard.tsx
