+++ front/src/pages/LanguagesPage/LanguagesPage.tsx
@@ -1,6 +1,6 @@
 import "./LanguagesPage.css";
 import { Container } from "../../components/Container/Container";
-import { ListCatalog } from "../../components/Catalogs/ListCatalog/ListCatalog";
+import { ListCatalog } from "../../components/ListCatalog/ListCatalog";
 import block from "bem-cn";
 import React from "react";
